chore(suite): split handling of connect modals and desktop api
