+++ packages/suite-web/src/MainWeb.tsx
@@ -17,6 +17,7 @@ import Resize from 'src/support/suite/Resize';
 import Protocol from 'src/support/suite/Protocol';
 import Autodetect from 'src/support/suite/Autodetect';
 import { useTor } from 'src/support/suite/useTor';
+import { useConnectPopupModals } from 'src/support/suite/useConnectPopupModals';
 import OnlineStatus from 'src/support/suite/OnlineStatus';
 import { ErrorBoundary } from 'src/support/suite/ErrorBoundary';
 import RouterHandler from 'src/support/suite/Router';
@@ -33,6 +34,7 @@ const MainWeb = () => {
     usePlaywright();
     useTor();
     useDebugLanguageShortcut();
+    useConnectPopupModals();
     const formattersConfig = useFormattersConfig();
 
     return (
